requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
yt-dlp==2024.5.27
python-dotenv==1.0.0
requests==2.32.3